requirements.txt
requests
BeautifulSoup4
certifi
cloudscraper
fake_useragent
python-dotenv
openai